Correct controler behaviour

+++ src/main/java/data/PeopleRepository.java
@@ -1,7 +1,7 @@
 package data;
 
-import org.springframework.context.annotation.Bean;
 import org.springframework.data.repository.CrudRepository;
+
 
 import java.util.List;
 
@@ -22,5 +22,5 @@ public interface PeopleRepository extends CrudRepository<People, Integer> {
 
     void delete(People people);
 
-    boolean exists(Long id);
+    boolean exists(int id);
 }
